Updated code to reflect changes in the MPFProj11 and MPFProj12 releases

+++ Dev10/Src/CSharp/ProjectDesignerDocumentManager.cs
@@ -91,9 +91,9 @@ namespace Microsoft.VisualStudio.Project
 
 				IVsUIHierarchy parentHierarchy = this.Node.ProjectManager.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchy) as IVsUIHierarchy;
 
-                int parentHierarchyItemId = (int)this.Node.ProjectManager.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchyItemid);
+                uint parentHierarchyItemId = (uint)this.Node.ProjectManager.GetProperty((int)__VSHPROPID.VSHPROPID_ParentHierarchyItemid);
 
-                ErrorHandler.ThrowOnFailure(uiShellOpenDocument.OpenSpecificEditor(editorFlags, fullPath, ref editorType, physicalView, ref logicalView, caption, parentHierarchy, (uint)parentHierarchyItemId, docDataExisting, serviceProvider, out frame));
+                ErrorHandler.ThrowOnFailure(uiShellOpenDocument.OpenSpecificEditor(editorFlags, fullPath, ref editorType, physicalView, ref logicalView, caption, parentHierarchy, parentHierarchyItemId, docDataExisting, serviceProvider, out frame));
 
 				if(frame != null)
 				{
